requirements.txt
# Core dependencies for SP500 Equity Data Pipeline
snowflake-connector-python>=3.6.0
requests>=2.31.0
pandas>=2.0.0
pyarrow>=14.0.0
boto3>=1.34.0

# Additional dependencies that may be required
botocore>=1.34.0
cryptography>=41.0.0
numpy>=1.24.0
